style: simplify conflict index documentation and move tests
