+++ memanto/cli/client/direct_client.py
@@ -43,6 +43,7 @@ from memanto.app.utils.errors import (
     SessionExpiredError,
     SessionNotFoundError,
 )
+from memanto.app.utils.temporal_helpers import utc_date_str
 from memanto.app.utils.validation import (
     InputLimits,
     is_successful_write_result,
@@ -1359,17 +1360,12 @@ class DirectClient:
             date: Date string (YYYY-MM-DD). Defaults to today.
 
         Returns:
-            List of unresolved conflict dicts. Each carries an ``index`` field
-            holding its STABLE position in the full report. ``resolve_conflict``
-            addresses conflicts by that full-report index, so callers must
-            resolve using this ``index`` value — NOT the position of the item
-            within this filtered list. The two diverge as soon as any earlier
-            conflict is resolved, and because resolution deletes memories,
-            resolving by filtered position would delete the wrong memory.
+            List of unresolved conflict dicts, each with a stable ``index``
+            into the full conflict report.
         """
 
         if not date:
-            date = datetime.now().strftime("%Y-%m-%d")
+            date = utc_date_str()
 
         json_path = (
             Path.home() / ".memanto" / "conflicts" / f"{agent_id}_{date}_conflicts.json"
@@ -1381,8 +1377,7 @@ class DirectClient:
         with open(json_path, encoding="utf-8") as f:
             all_conflicts = json.load(f)
 
-        # Return only unresolved conflicts, each tagged with its stable index
-        # into the full report so callers can resolve them unambiguously.
+        # Keep unresolved conflicts but preserve each full-report index.
         return [
             {**c, "index": idx}
             for idx, c in enumerate(all_conflicts)
@@ -1404,11 +1399,8 @@ class DirectClient:
         Args:
             agent_id: Target agent.
             date: Date string (YYYY-MM-DD).
-            conflict_index: Stable 0-based index into the FULL conflict report,
-                as returned in the ``index`` field of each item from
-                ``list_conflicts``. Do not pass the position of the item within
-                the filtered ``list_conflicts`` result — those diverge once any
-                earlier conflict is resolved.
+            conflict_index: Stable 0-based index into the full conflict report
+                (use ``list_conflicts(...)[i]["index"]``).
             action: Resolution action — ``keep_old``, ``keep_new``,
                 ``keep_both``, ``remove_both``, or ``manual``.
             manual_content: Required when action is ``manual``.
@@ -1440,11 +1432,7 @@ class DirectClient:
 
         conflict = all_conflicts[conflict_index]
 
-        # Defense in depth against a stale / desynced index: refuse to act on a
-        # conflict that is already resolved. Without this, a caller that resolved
-        # by filtered-list position (instead of the stable ``index`` from
-        # list_conflicts) could silently re-run a resolution and delete a memory
-        # that was never the one they selected.
+        # Guard against stale/desynced conflict indexes.
         if conflict.get("resolved", False):
             raise ValueError(
                 f"Conflict at index {conflict_index} is already resolved. "
