+++ tests/test_cli.py
@@ -15,6 +15,8 @@ import pytest
 from typer.testing import CliRunner
 
 from memanto.app.clients.backend import Backend
+from memanto.cli.client import direct_client as direct_client_module
+from memanto.cli.client.direct_client import DirectClient
 from memanto.cli.main import app
 
 runner = CliRunner()
@@ -120,6 +122,45 @@ def mock_all_clients():
 
     for p in patches:
         p.stop()
+
+
+def _conflict(letter: str) -> dict:
+    return {
+        "type": "contradiction",
+        "title": f"Conflict {letter}",
+        "old_memory_id": f"{letter}_old",
+        "old_content": f"old {letter}",
+        "new_memory_id": f"{letter}_new",
+        "new_content": f"new {letter}",
+        "description": f"conflict {letter}",
+        "recommendation": "keep_new",
+        "resolved": False,
+        "resolution": None,
+    }
+
+
+def _write_conflict_report(
+    tmp_path, agent_id: str, date: str, conflicts: list[dict]
+) -> None:
+    path = tmp_path / ".memanto" / "conflicts" / f"{agent_id}_{date}_conflicts.json"
+    path.parent.mkdir(parents=True, exist_ok=True)
+    path.write_text(json.dumps(conflicts), encoding="utf-8")
+
+
+def _make_direct_client(tmp_path, monkeypatch):
+    monkeypatch.setattr(
+        direct_client_module.Path, "home", classmethod(lambda cls: tmp_path)
+    )
+    client = DirectClient(api_key="test-key")
+    mock_write = MagicMock()
+    mock_write.delete_memory.return_value = True
+    mock_write.store_memory.return_value = {"id": "manual_new"}
+    client._write_service = mock_write
+    return client, mock_write
+
+
+def _deleted_ids(mock_write) -> list[str]:
+    return [call.args[0] for call in mock_write.delete_memory.call_args_list]
 
 
 class TestMEMANTOCLI:
@@ -1124,9 +1165,19 @@ class TestMEMANTOCLI:
         mock_all_clients.generate_daily_summary.return_value = {
             "summary": {"status": "success", "summary_path": "summary.md"},
         }
-        result = runner.invoke(app, ["daily-summary"])
+        with patch(
+            "memanto.cli.commands.memory.utc_date_str",
+            return_value="2026-07-30",
+        ):
+            result = runner.invoke(app, ["daily-summary"])
+
         assert result.exit_code == 0
         assert "generated" in result.stdout.lower()
+        mock_all_clients.generate_daily_summary.assert_called_once_with(
+            agent_id="test-agent",
+            date="2026-07-30",
+            output_path=None,
+        )
 
     def test_detect_conflicts(self, mock_all_clients):
         """Test 'memanto detect-conflicts'"""
@@ -1137,9 +1188,18 @@ class TestMEMANTOCLI:
                 "json_path": "conflicts.json",
             },
         }
-        result = runner.invoke(app, ["detect-conflicts"])
+        with patch(
+            "memanto.cli.commands.memory.utc_date_str",
+            return_value="2026-07-30",
+        ):
+            result = runner.invoke(app, ["detect-conflicts"])
+
         assert result.exit_code == 0
         assert "conflict report generated" in result.stdout.lower()
+        mock_all_clients.generate_conflict_report.assert_called_once_with(
+            agent_id="test-agent",
+            date="2026-07-30",
+        )
 
     def test_conflicts_list(self, mock_all_clients):
         """Test 'memanto conflicts --list'"""
@@ -1152,9 +1212,93 @@ class TestMEMANTOCLI:
                 "recommendation": "merge",
             }
         ]
-        result = runner.invoke(app, ["conflicts", "--list"])
+        with patch(
+            "memanto.cli.commands.memory.utc_date_str",
+            return_value="2026-07-30",
+        ):
+            result = runner.invoke(app, ["conflicts", "--list"])
+
         assert result.exit_code == 0
         assert "Found 1 unresolved conflict" in result.stdout
+        mock_all_clients.list_conflicts.assert_called_once_with(
+            agent_id="test-agent",
+            date="2026-07-30",
+        )
+
+    def test_list_conflicts_exposes_stable_full_report_index(self, tmp_path, monkeypatch):
+        """Unresolved conflicts keep their original full-report index."""
+        agent_id, date = "agent-1", "2026-07-01"
+        _write_conflict_report(
+            tmp_path, agent_id, date, [_conflict("A"), _conflict("B"), _conflict("C")]
+        )
+        client, _ = _make_direct_client(tmp_path, monkeypatch)
+
+        listed = client.list_conflicts(agent_id=agent_id, date=date)
+        assert [(c["title"], c["index"]) for c in listed] == [
+            ("Conflict A", 0),
+            ("Conflict B", 1),
+            ("Conflict C", 2),
+        ]
+
+        client.resolve_conflict(agent_id, date, conflict_index=0, action="keep_new")
+        listed_after = client.list_conflicts(agent_id=agent_id, date=date)
+        assert [(c["title"], c["index"]) for c in listed_after] == [
+            ("Conflict B", 1),
+            ("Conflict C", 2),
+        ]
+
+    def test_resolving_by_provided_index_deletes_the_correct_memory(
+        self, tmp_path, monkeypatch
+    ):
+        """Resolve by stable index so only the selected memory is deleted."""
+        agent_id, date = "agent-1", "2026-07-01"
+        _write_conflict_report(
+            tmp_path, agent_id, date, [_conflict("A"), _conflict("B"), _conflict("C")]
+        )
+        client, mock_write = _make_direct_client(tmp_path, monkeypatch)
+
+        a = next(
+            c
+            for c in client.list_conflicts(agent_id=agent_id, date=date)
+            if c["title"] == "Conflict A"
+        )
+        client.resolve_conflict(
+            agent_id, date, conflict_index=a["index"], action="keep_new"
+        )
+        assert _deleted_ids(mock_write) == ["A_old"]
+
+        remaining = client.list_conflicts(agent_id=agent_id, date=date)
+        c = next(x for x in remaining if x["title"] == "Conflict C")
+        client.resolve_conflict(
+            agent_id, date, conflict_index=c["index"], action="keep_new"
+        )
+
+        assert _deleted_ids(mock_write)[1:] == ["C_old"]
+
+        report = json.loads(
+            (
+                tmp_path / ".memanto" / "conflicts" / f"{agent_id}_{date}_conflicts.json"
+            ).read_text(encoding="utf-8")
+        )
+        by_title = {row["title"]: row for row in report}
+        assert by_title["Conflict C"]["resolved"] is True
+        assert by_title["Conflict B"]["resolved"] is False
+
+    def test_resolving_an_already_resolved_index_is_rejected(
+        self, tmp_path, monkeypatch
+    ):
+        """Resolved conflicts cannot be resolved again by stale indexes."""
+        agent_id, date = "agent-1", "2026-07-01"
+        _write_conflict_report(tmp_path, agent_id, date, [_conflict("A"), _conflict("B")])
+        client, mock_write = _make_direct_client(tmp_path, monkeypatch)
+
+        client.resolve_conflict(agent_id, date, conflict_index=0, action="keep_new")
+        assert _deleted_ids(mock_write) == ["A_old"]
+
+        with pytest.raises(ValueError, match="already resolved"):
+            client.resolve_conflict(agent_id, date, conflict_index=0, action="keep_new")
+
+        assert _deleted_ids(mock_write) == ["A_old"]
 
     def test_memory_export(self, mock_all_clients):
         """Test 'memanto memory export'"""
